requirements.txt
rdkit>=2023.9.1
biopython>=1.81
prody>=2.4.1
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
propka>=3.5.0
pdb2pqr>=3.6.0
pytest>=7.4.0
pyyaml>=6.0
requests>=2.31.0
dataclasses>=0.6; python_version < "3.7"